# packages/api-client/specs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/knowledge/create": {
      "post": {
        "operationId": "KnowledgeController_createKnowledge",
        "summary": "创建知识点",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateKnowledgeDto"
              },
              "examples": {
                "example1": {
                  "value": {
                    "title": "谚语",
                    "content": "不到长城非好汉"
                  }
                },
                "example2": {
                  "value": {
                    "content": "title 可以为空"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "知识点管理"
        ]
      }
    }
  },
  "info": {
    "title": "背书记忆 API",
    "description": "Agent 背书记忆项目接口文档",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateKnowledgeDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "example": "不到长城非好汉",
            "description": "知识点名称"
          }
        },
        "required": [
          "content"
        ]
      }
    }
  }
}