# packages/api-client/src/generated/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export type ClientOptions = {
    baseUrl: string;
};

export type SendOtpDto = {
    /**
     * 邮箱地址
     */
    email: string;
};

export type VerifyOtpDto = {
    /**
     * 邮箱地址
     */
    email: string;
    /**
     * 验证码
     */
    code: string;
};

export type CreateKnowledgeDto = {
    /**
     * 知识点名称
     */
    content: string;
};

export type HealthControllerCheckHealthData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/health';
};

export type HealthControllerCheckHealthResponses = {
    /**
     * 服务正常
     */
    200: unknown;
};

export type HealthControllerPingData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/health/ping';
};

export type HealthControllerPingResponses = {
    /**
     * 返回 pong
     */
    200: unknown;
};

export type HealthControllerGetServerTimeData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/health/authTime';
};

export type HealthControllerGetServerTimeResponses = {
    /**
     * 返回服务器时间
     */
    200: unknown;
};

export type AuthControllerSendOtpData = {
    body: SendOtpDto;
    path?: never;
    query?: never;
    url: '/auth/send-otp';
};

export type AuthControllerSendOtpResponses = {
    /**
     * 返回验证码
     */
    200: unknown;
};

export type AuthControllerVerifyOtpData = {
    body: VerifyOtpDto;
    path?: never;
    query?: never;
    url: '/auth/verify-otp';
};

export type AuthControllerVerifyOtpResponses = {
    /**
     * 返回验证结果
     */
    200: unknown;
};

export type KnowledgeControllerGetKnowledgeListData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/knowledge';
};

export type KnowledgeControllerGetKnowledgeListResponses = {
    201: unknown;
};

export type KnowledgeControllerCreateKnowledgeData = {
    body: CreateKnowledgeDto;
    path?: never;
    query?: never;
    url: '/knowledge/create';
};

export type KnowledgeControllerCreateKnowledgeResponses = {
    201: unknown;
};